test: extract native darwin test to main thread
